Namespacing the PPMS client
