+++ config/initializers/breakers.rb
@@ -19,7 +19,7 @@ require 'evss/gi_bill_status/service'
 require 'evss/pciu_address/configuration'
 require 'evss/reference_data/configuration'
 require 'facilities/bulk_configuration'
-require 'facilities/ppms/configuration'
+require 'facilities/ppms/v0/configuration'
 require 'gi/configuration'
 require 'gibft/configuration'
 require 'hca/configuration'
@@ -59,7 +59,7 @@ services = [
   Gibft::Configuration.instance.breakers_service,
   Facilities::AccessWaitTimeConfiguration.instance.breakers_service,
   Facilities::AccessSatisfactionConfiguration.instance.breakers_service,
-  Facilities::PPMS::Configuration.instance.breakers_service,
+  Facilities::PPMS::V0::Configuration.instance.breakers_service,
   GI::Configuration.instance.breakers_service,
   HCA::Configuration.instance.breakers_service,
   MHVAC::Configuration.instance.breakers_service,
