+++ spec/lib/facilities/ppms/v0/response_spec.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
 require 'rails_helper'
-require 'facilities/ppms/response'
+require 'facilities/ppms/v0/response'
 
-describe Facilities::PPMS::Response, team: :facilities do
+describe Facilities::PPMS::V0::Response, team: :facilities do
   describe '.from_provider_locator' do
     it 'creates Providers from a ppms response' do
     end
